pyproject.toml
[tool.poetry]
name = "repositories-api"
version = "0.1.0"
description = ""
authors = ["Marcio Gualtieri <marcio.gualtieri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
gidgethub = "^5.3.0"
cachetools = "^5.3.3"
aiohttp = "^3.9.5"
pydantic-settings = "^2.2.1"
uvicorn = "^0.29.0"

[tool.poetry.dev-dependencies]
vcrpy = "^6.0.1"
pytest = "^8.2.0"
honcho = "^1.1.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
pylint = "^3.1.0"
coverage = "^7.5.1"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = [
    "vcr"
]
ignore_missing_imports = true
disallow_untyped_defs = true

